@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestRange(t *testing.T) {
@@ -19,13 +18,12 @@ func TestRange(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run(fmt.Sprintf("[%v, %v)", c.min, c.max), func(t *testing.T) {
-			r := Range(c.min, c.max)
-			for i := c.min; i < c.max; i++ {
-				ok := r.Next()
-				require.True(t, ok)
-				assert.Equal(t, i, r.Value())
+			i := c.min
+			for v := range Range(c.min, c.max) {
+				assert.Equal(t, i, v)
+				i++
 			}
-			assert.False(t, r.Next())
+			assert.Equal(t, c.max, i)
 		})
 	}
 }
@@ -41,11 +39,13 @@ func TestMinRange(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run(fmt.Sprintf("[%v, ...)", c.min), func(t *testing.T) {
-			r := MinRange(c.min)
-			for i, j := c.min, 0; j < 10; i, j = i+1, j+1 {
-				ok := r.Next()
-				require.True(t, ok)
-				assert.Equal(t, i, r.Value())
+			i, j := c.min, 0
+			for v := range MinRange(c.min) {
+				if j == 10 {
+					return
+				}
+				assert.Equal(t, i, v)
+				i, j = i+1, j+1
 			}
 		})
 	}
