@@ -1,8 +1,10 @@
 package fx
 
-func Reduce[T, U any](it Iterator[T], init U, fn func(acc U, v T) U) U {
-	for it.Next() {
-		init = fn(init, it.Value())
+import "iter"
+
+func Reduce[T, U any](it iter.Seq[T], init U, fn func(acc U, v T) U) U {
+	for v := range it {
+		init = fn(init, v)
 	}
 	return init
 }
