@@ -1,8 +1,11 @@
 package fx
 
-import "sort"
+import (
+	"iter"
+	"sort"
+)
 
-func Sorted[T any](it Iterator[T], less func(a, b T) bool) Iterator[T] {
+func Sorted[T any](it iter.Seq[T], less func(a, b T) bool) iter.Seq[T] {
 	s := ToSlice(it)
 	sort.Slice(s, func(i, j int) bool { return less(s[i], s[j]) })
 	return IterSlice(s)
